FilterAccessor: move to new module
